Move CombinePartialFinalAggregate rule into physical-optimizer crate
