+++ datafusion/core/tests/physical_optimizer/mod.rs
@@ -16,6 +16,7 @@
 // under the License.
 
 mod aggregate_statistics;
+mod combine_partial_final_agg;
 mod limit_pushdown;
 mod limited_distinct_aggregation;
 mod test_util;
